Разнес код по разным модулям
